+++ R/phylo.R
@@ -318,6 +318,8 @@ AddTip <- function (tree,
 #' @export
 AddTipEverywhere <- function (tree, label = 'New tip', includeRoot = FALSE) {
   nTip <- NTip(tree)
+  if (nTip == 0L) return(list(SingleTaxonTree(label)))
+  if (nTip == 1L) return(list(StarTree(c(tree$tip.label, label))))
   whichNodes <- seq_len(nTip + tree$Nnode)
   edge <- tree$edge
   root <- RootNode(edge)
@@ -326,7 +328,7 @@ AddTipEverywhere <- function (tree, label = 'New tip', includeRoot = FALSE) {
     child <- edge[, 2]
     rootChildren <- child[parent == root]
 
-    whichNodes <- if (length(rootChildren) == 2L) {
+    whichNodes <- if (length(rootChildren) == 2L && nTip > 2L) {
       rootChildrenNodes <- rootChildren[rootChildren > nTip]
       whichNodes[-c(root, rootChildrenNodes[1])]
     } else {
@@ -459,7 +461,7 @@ CladeSizes <- function (tree, internal = FALSE, nodes = NULL) {
     internal <- isTRUE(internal)
   }
 
-  edge <- Postorder(tree$edge, renumber = FALSE)
+  edge <- Postorder(tree$edge, renumber = FALSE, sizeSort = TRUE)
   nTip <- NTip(tree)
 
   size <- c(rep.int(1L, nTip), rep.int(internal, max(edge[, 1]) - nTip))