+++ src/RcppExports.cpp
@@ -6,6 +6,11 @@
 
 using namespace Rcpp;
 
+#ifdef RCPP_USE_GLOBAL_ROSTREAM
+Rcpp::Rostream<true>&  Rcpp::Rcout = Rcpp::Rcpp_cout_get();
+Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
+#endif
+
 // as_newick
 CharacterVector as_newick(IntegerMatrix edge);
 RcppExport SEXP _TreeTools_as_newick(SEXP edgeSEXP) {
@@ -14,6 +19,30 @@ BEGIN_RCPP
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< IntegerMatrix >::type edge(edgeSEXP);
     rcpp_result_gen = Rcpp::wrap(as_newick(edge));
+    return rcpp_result_gen;
+END_RCPP
+}
+// consensus_tree
+LogicalMatrix consensus_tree(const List trees, const NumericVector p);
+RcppExport SEXP _TreeTools_consensus_tree(SEXP treesSEXP, SEXP pSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const List >::type trees(treesSEXP);
+    Rcpp::traits::input_parameter< const NumericVector >::type p(pSEXP);
+    rcpp_result_gen = Rcpp::wrap(consensus_tree(trees, p));
+    return rcpp_result_gen;
+END_RCPP
+}
+// drop_tip
+IntegerMatrix drop_tip(const IntegerMatrix edge, const IntegerVector drop);
+RcppExport SEXP _TreeTools_drop_tip(SEXP edgeSEXP, SEXP dropSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const IntegerMatrix >::type edge(edgeSEXP);
+    Rcpp::traits::input_parameter< const IntegerVector >::type drop(dropSEXP);
+    rcpp_result_gen = Rcpp::wrap(drop_tip(edge, drop));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -74,6 +103,18 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< IntegerMatrix >::type edge(edgeSEXP);
     Rcpp::traits::input_parameter< IntegerVector >::type nTip(nTipSEXP);
     rcpp_result_gen = Rcpp::wrap(cpp_edge_to_splits(edge, nTip));
+    return rcpp_result_gen;
+END_RCPP
+}
+// splits_to_edge
+IntegerMatrix splits_to_edge(const RawMatrix splits, const IntegerVector nTip);
+RcppExport SEXP _TreeTools_splits_to_edge(SEXP splitsSEXP, SEXP nTipSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const RawMatrix >::type splits(splitsSEXP);
+    Rcpp::traits::input_parameter< const IntegerVector >::type nTip(nTipSEXP);
+    rcpp_result_gen = Rcpp::wrap(splits_to_edge(splits, nTip));
     return rcpp_result_gen;
 END_RCPP
 }