+++ tests/testthat/test-Support.R
@@ -1,12 +1,10 @@
-context("Support.R")
-
 test_that("Node supports calculated correctly", {
   treeSample <- list(
     correct = ape::read.tree(text = "((((((A,B),C),D),E),F),out);"),
     swapFE  = ape::read.tree(text = "((((((A,B),C),D),F),E),out);"),
     DEClade = ape::read.tree(text = "(((((A,B),C),(D,E)),F),out);"),
     swapBC  = ape::read.tree(text = "((((((A,C),B),D),E),F),out);"),
-    DbyA    = ape::read.tree(text = "((((((A,D),C),B),E),F),out);")
+    DbyA    = ape::read.tree(text = "((((((A,D),C),B),E),F,G),out);")
   )
   expect_equal(c('10' = 4, '11' = 4, '12' = 4, '13' = 3),
                SplitFrequency(treeSample$correct, treeSample))
@@ -34,14 +32,21 @@ test_that("SplitFrequency() handles four-split trees", {
 })
 
 test_that("LabelSplits()", {
-  SplitLabelling <- function () {
+  expect_error(LabelSplits(BalancedTree(8), 1:8))
+  skip_if_not_installed('vdiffr', minimum_version = "1.0.0")
+  skip_if(packageVersion("graphics") < "4.1.0")
+  vdiffr::expect_doppelganger('LabelSplits()',  function () {
+    tree <- BalancedTree(9)
+    plot(tree)
+    labs <- letters[6:1]
+    names(labs) <- rev(names(as.Splits(tree)))
+    LabelSplits(tree, labs, frame = 'circ', cex = 2, bg = 'orange')
+  })
+  vdiffr::expect_doppelganger('LabelSplits()-names', function() {
     tree <- BalancedTree(9)
     plot(tree)
     labs <- letters[1:6]
-    names(labs) <- names(as.Splits(tree))
-    LabelSplits(tree, labs, frame = 'circ', cex = 2, bg = 'orange')
-  }
-  skip_if_not_installed('vdiffr', minimum_version = "1.0.0")
-  skip_if(packageVersion("graphics") < "4.1.0")
-  # vdiffr::expect_doppelganger('LabelSplits()', SplitLabelling)
+    LabelSplits(tree, labs, bg = 'orange')
+    expect_warning(LabelSplits(BalancedTree(9), setNames(letters[11:16], 1:6)))
+  })
 })