+++ tests/testthat/test-mst.R
@@ -1,5 +1,3 @@
-context('mst.R')
-
 test_that('MST edges calculated correctly', {
   set.seed(0)
   points <- matrix(c(0.1, 0, 1.9, 2, 1.1, 1,
@@ -19,7 +17,7 @@ test_that('MST edges calculated correctly', {
   }
   skip_if_not_installed('vdiffr', minimum_version = "1.0.0")
   skip_if(packageVersion("graphics") < "4.1.0")
-  # vdiffr::expect_doppelganger('MST plotting', MSTPlot)
+  vdiffr::expect_doppelganger('MST plotting', MSTPlot)
 })
 test_that("MST handles large distance matrices", {
   x <- dist(0:300)