+++ tests/testthat/test-tree_generation.R
@@ -1,5 +1,3 @@
-context('tree_generation.R')
-
 test_that('Pectinate trees are generated', {
   expect_equal(ape::read.tree(text = '(t1, (t2, (t3, t4)));'),
                PectinateTree(4L))
@@ -21,7 +19,8 @@ test_that('Balanced trees are generated correctly', {
                BalancedTree(9L))
   expect_equal(BalancedTree(as.character(1:9)), BalancedTree(1:9))
   escapees <- c("Apostrophe's", 'and quote"s')
-  expect_equivalent(PectinateTree(escapees), BalancedTree(escapees))
+  expect_equal(ignore_attr = TRUE,
+               PectinateTree(escapees), BalancedTree(escapees))
   expect_equal(integer(0), .BalancedBit(seq_len(0)))
   expect_equal('Test', .BalancedBit('Test'))
   expect_true(is.integer(BalancedTree(8)$edge))
@@ -35,7 +34,8 @@ test_that("StarTree() works", {
 
 test_that("Random trees are generated correctly", {
   expect_equal(c(4, 4, 5, 5, 1, 5, 2, 3), RandomTree(3, root = TRUE)$edge[1:8])
-  expect_equal(PectinateTree(c('t2', 't3', 't1')), RandomTree(3, root = 't2'))
+  expect_true(all.equal(RandomTree(3, root = 't2'),
+    PectinateTree(c('t2', 't3', 't1'))))
   expect_equal(c(4, 4, 4), RandomTree(3, root = FALSE)$edge[1:3])
   expect_warning(expect_equal(RandomTree(3, root = 't2'),
                               RandomTree(3, root = 2:3)))
@@ -49,16 +49,39 @@ test_that("NJTree() works", {
   a..f <- letters[1:6]
   bal6 <- StringToPhyDat('111100 111000 111000 110000', letters[1:6],
                          byTaxon = FALSE)
-  expect_equal(BalancedTree(letters[c(1:3, 6:4)]), RootTree(NJTree(bal6), a..f[1:3]))
+  expect_true(all.equal(
+    RootTree(NJTree(bal6), a..f[1:3]),
+    BalancedTree(letters[c(1:3, 6:4)])
+  ))
   expect_equal(c(0, 1, 2, 1, rep(0, 6)),
                Preorder(NJTree(bal6, TRUE))$edge.length * 4L)
+})
+
+test_that("Constrained NJ trees work", {
+  dataset <- MatrixToPhyDat(matrix(
+    c(0, 1, 1, 1, 0, 1,
+      0, 1, 1, 0, 0, 1), ncol = 2,
+    dimnames = list(letters[1:6], NULL)))
+  constraint <- MatrixToPhyDat(c(a = 0, b = 0, c = 0, d = 0, e = 1, f = 1))
+  expect_true(all.equal(read.tree(text = "(a, (d, ((c, b), (e, f))));"),
+                        ConstrainedNJ(dataset, constraint)))
+  # b == c == f, so these three could be resolved in one of three ways. Drop B.
+  expect_true(all.equal(DropTip(NJTree(dataset), 'b'),
+                        DropTip(ConstrainedNJ(dataset, dataset), 'b')))
+
+  expect_true(all.equal(
+    KeepTip(ConstrainedNJ(dataset, constraint[3:6]), letters[3:6]),
+    BalancedTree(letters[3:6])
+  ))
 })
 
 test_that("EnforceOutgroup() fails nicely", {
   expect_error(EnforceOutgroup(BalancedTree(6), 'Non-taxon'))
   expect_error(EnforceOutgroup(BalancedTree(6), c('t1', 'Non-taxon')))
-  expect_equal(BalancedTree(letters[5:6]),
-               Subtree(Preorder(EnforceOutgroup(letters[1:8], letters[5:6])), 15))
+  expect_true(all.equal(
+    BalancedTree(letters[5:6]),
+    Subtree(Preorder(EnforceOutgroup(letters[1:8], letters[5:6])), 15)
+    ))
   expect_equal(ape::root(BalancedTree(8), 't1', resolve.root = TRUE),
                EnforceOutgroup(BalancedTree(8), 't1'))
 })